+++ docs/change-log.md
@@ -27,7 +27,8 @@ This file records the major version changes of the projects.
 - initialised `client/` for the frontend building, empowered by tailwind-CSS and React.
 - created data flow and UML diagram to outline the project and database structure.
 - created user interfaces and basic CRUD functionalities for manage candidates page.
-- created sorting methods and user interface for manage candidate pages.
+- created sorting methods and user interface for manage candidates page.
+- created filtering methods and user interface for manage candidates page.
 
 ### Edited